Adapt to `macaw-symbolic` introducing the `MemModelConfig` data type
https://github.com/GaloisInc/macaw/pull/327 changed the types of some
`macaw-symbolic` API functions to consolidate certain memory model–related
configuration options into the `MemModelConfig` data type. This patch makes the
necessary changes on the `stubs-common` side.

https://github.com/GaloisInc/macaw/pull/327 also introduces a lazy memory
model, and if we put in more elbow grease, we could potentially remove much of
`stubs-common`'s memory model by building on top of that. This would be
significantly more work, however, so we defer this task to
https://github.com/GaloisInc/stubs/issues/12.