Bump macaw submodule to bring in changes from GaloisInc/macaw#401
This bumps the `macaw` submodule to bring in the changes from
GaloisInc/macaw#401, which is a prerequisite to being able to support PowerPC
in `stubs`. This also requires bumping the `crucible`, `elf-edit`,
`llvm-pretty`, and `what4` submodules in order to construct a build plan that
is consistent with the latest `macaw` commit.

One step towards #33.